Move BaseOpHints to API module
