+++ scijava/scijava-ops-api/src/main/java/org/scijava/ops/api/BaseOpHints.java
@@ -1,4 +1,4 @@
-package org.scijava.ops.engine;
+package org.scijava.ops.api;
 
 import java.lang.annotation.Annotation;
