+++ scijava/scijava-ops-engine/src/main/java/org/scijava/ops/engine/hint/AdaptationHints.java
@@ -6,7 +6,7 @@ import java.util.Map;
 import java.util.UUID;
 
 import org.scijava.ops.api.Hints;
-import org.scijava.ops.engine.BaseOpHints.Adaptation;
+import org.scijava.ops.api.BaseOpHints.Adaptation;
 
 /**
  * A set of {@link Hints} governing Adaptation procedures.
