+++ scijava/scijava-ops-engine/src/main/java/org/scijava/ops/engine/impl/DefaultOpEnvironment.java
@@ -58,9 +58,9 @@ import org.scijava.ops.api.OpInfo;
 import org.scijava.ops.api.OpInfoGenerator;
 import org.scijava.ops.api.OpRef;
 import org.scijava.ops.api.OpWrapper;
-import org.scijava.ops.engine.BaseOpHints.Adaptation;
-import org.scijava.ops.engine.BaseOpHints.DependencyMatching;
-import org.scijava.ops.engine.BaseOpHints.Simplification;
+import org.scijava.ops.api.BaseOpHints.Adaptation;
+import org.scijava.ops.api.BaseOpHints.DependencyMatching;
+import org.scijava.ops.api.BaseOpHints.Simplification;
 import org.scijava.ops.engine.OpHistoryService;
 import org.scijava.ops.engine.OpInstance;
 import org.scijava.ops.engine.hint.AdaptationHints;
