+++ scijava/scijava-ops-engine/src/main/java/org/scijava/ops/engine/simplify/PrimitiveSimplifiers.java
@@ -4,7 +4,7 @@ package org.scijava.ops.engine.simplify;
 import java.util.function.Function;
 
 import org.scijava.ops.api.OpHints;
-import org.scijava.ops.engine.BaseOpHints.Simplification;
+import org.scijava.ops.api.BaseOpHints.Simplification;
 import org.scijava.ops.spi.OpCollection;
 import org.scijava.ops.spi.OpField;
 import org.scijava.plugin.Plugin;
