+++ scijava/scijava-ops-engine/src/test/java/org/scijava/ops/engine/simplify/SimplificationPriorityTest.java
@@ -7,8 +7,8 @@ import java.util.function.Function;
 import org.junit.Assert;
 import org.junit.Test;
 import org.scijava.ops.api.OpHints;
+import org.scijava.ops.api.BaseOpHints.Simplification;
 import org.scijava.ops.engine.AbstractTestEnvironment;
-import org.scijava.ops.engine.BaseOpHints.Simplification;
 import org.scijava.ops.engine.conversionLoss.LossReporter;
 import org.scijava.ops.spi.OpCollection;
 import org.scijava.ops.spi.OpField;
